% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_INE.R
\docType{data}
\name{cod_INE_muni_16}
\alias{cod_INE_muni_16}
\title{Codigos INE municipios 2016}
\format{A data frame with 8125 rows (1 municipio mas que en 2017) and 6 variables :
\itemize{
\item CODAUTO: Código de CCAA
\item CPRO: Código de la provincia
\item CMUN: Código del municipio
\item DC: Código de control
\item NOMBRE: Nombre del municipio
\item INECodMuni: CPRO+CMUN (5 digits code)
}}
\source{
\url{http://www.ine.es}
}
\usage{
cod_INE_muni_16
}
\description{
Codigos del INE para municipios españoles a 1 enero 2016: 8.125 municipios, 6 variables.
Un municipio más que en 2017
}
\keyword{datasets}
