% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_INE.R
\docType{data}
\name{cod_INE_muni_17}
\alias{cod_INE_muni_17}
\title{Codigos INE municipios 2017}
\format{A data frame with 8124 rows and 4 variables: Cerdedo-Cotobade se unieron en 2016
\itemize{
\item CODAUTO: Código de CCAA
\item CPRO: Código de la provincia
\item CMUN: Código del municipio
\item DC: Código de control
\item NOMBRE: Nombre del municipio
\item INECodMuni: CPRO+CMUN (5 digits code)
}}
\source{
\url{http://www.ine.es/daco/daco42/codmun/codmunmapa.htm}
}
\usage{
cod_INE_muni_17
}
\description{
Codigos del INE para municipios españoles a 1 enero 2017: 8.124 municipios, 6 variables.
}
\keyword{datasets}
