% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_INE.R
\docType{data}
\name{cod_INE_prov}
\alias{cod_INE_prov}
\title{Codigos INE para provincias y CC.AA. (SPAIN)}
\format{A data frame with 52 rows and 4 variables:
\itemize{
\item CODAUTO: Código de la C.A.
\item CCAA: Nombre de la C.A.
\item CPRO: Código de la provincia.
\item PROVINCIA: Nombre de la provincia.
}}
\source{
\url{http://www.ine.es}
\url{http://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736177031&menu=ultiDatos&idp=1254734710990}
}
\usage{
cod_INE_prov
}
\description{
Codigos y nombres oficiales de las provincias y CCAA (SPAIN). Sacados del INE.
52 provincias (51 = Ceuta y 52 = Melilla).
19 CC.AA (18 = Ceuta y 19 = Melilla).
}
\examples{
\dontrun{
 df <- cod_INE_prov
}

}
\keyword{datasets}
