% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGN_data_lindes.R
\docType{data}
\name{IGN_CCAA_17}
\alias{IGN_CCAA_17}
\title{LINDES del IGN para CC.AA (SPAIN)}
\format{A sf-data frame with 19 rows and 11 variables:
\itemize{
\item INECodCCAA: Código INE de la C.A (Galicia = 12)
\item NATCODE: Cogigo del IGN (Galicia = 34120000000)
\item NAMEUNIT: Nombre de la CC.AA (IGN)
\item NombreCCAA: Nombre de la CC.AA (INE)
\item INSPIREID: IGN code (Galicia = ES.IGN.SIGLIM34120000000)
\item COUNTRY: Country code (ES = Spain)
\item NATLEVEL: http://inspire.ec.europa.eu/codelist/AdministrativeHierarchyLevel/2ndOrder
\item CODNUT1: (Galicia = ES1)
\item CODNUT2: (Galicia = ES1)
\item CODNUT3: NA (no hay datos)
\item geometry: la geometría
}}
\source{
\url{http://centrodedescargas.cnig.es/CentroDescargas/index.jsp}
}
\usage{
IGN_CCAA_17
}
\description{
Poligonos/geometrias/shapes de las provincias  CC.AA. (SPAIN). Sacados del IGN
19 CC.AA (18 = Ceuta y 19 = Melilla).
}
\examples{
\dontrun{
 df <- IGN_CCAA_17
}

}
\keyword{datasets}
