% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGN_data_lindes.R
\docType{data}
\name{IGN_CCAA_17s}
\alias{IGN_CCAA_17s}
\title{LINDES del IGN para CC.AA (SPAIN) (version short)}
\format{A sf-data frame with 19 rows and 4 variables:
\itemize{
\item INECodCCAA: Código INE de la C.A (Galicia = 12)
\item NATCODE: Cogigo del IGN (Galicia = 34120000000)
\item NombreCCAA: Nombre de la CC.AA (INE)
\item geometry: la geometría
}}
\source{
\url{http://centrodedescargas.cnig.es/CentroDescargas/index.jsp}
}
\usage{
IGN_CCAA_17s
}
\description{
Poligonos/geometrias/shapes de las provincias  CC.AA. (SPAIN). Sacados del IGN
19 CC.AA (18 = Ceuta y 19 = Melilla). Solo 4 variables
}
\examples{
\dontrun{
 df <- IGN_CCAA_17s
}

}
\keyword{datasets}
