% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGN_data_lindes.R
\docType{data}
\name{IGN_mun_17}
\alias{IGN_mun_17}
\title{LINDES del IGN para Municipios (SPAIN)}
\format{A data frame with 8211 rows and 20 variables:
\itemize{
\item INECodMuni: Codigo INE municipio (5 digitos (Pancrudo = 44177))
\item NATCODE: Codigo IGN (Pancrudo = 34024444177)
\item NAMEUNIT: Nombre municipio (IGN)
\item NombreMuni: Nombre municipio (INE)
\item NombreProv: Nombre provincia (INE)
\item NombreCCAA: Nombre de la C.A (INE)
\item INSPIREID: (Pancrudo = ES.IGN.SIGLIM34024444177)
\item COUNTRY: (ES = Spain)
\item NATLEVEL: http://inspire.ec.europa.eu/codelist/AdministrativeHierarchyLevel/4thOrder
\item CODNUT1: (Pancrudo = ES2)
\item CODNUT2: (Pancrudo = ES24)
\item CODNUT3: (Pancrudo = ES242)
\item INECodCCAA: Codigo CCAA(INE) (Pancrudo = 02)
\item INECodProv: Codigo provincial(INE)(Pancrudo = 44)
\item CMUN: Código municipal(INE)(Pancrudo = 177)
\item DC: Digito control (INE) (pancrudo = 4)
\item INENumMuni: Código municipal(INE) (numeric)(Pancrudo = 177)
\item INENumProv: Codigo provincial(INE) (numeric)(Pancrudo = 44)
\item INENumCCAA: Codigo CCAA(INE) (numeric)(Pancrudo = 02)
\item geometry: la geometria
}}
\source{
\url{http://centrodedescargas.cnig.es/CentroDescargas/index.jsp}
}
\usage{
IGN_mun_17
}
\description{
Poligonos/geometrias/shapes de los municipios  CC.AA. (SPAIN). Sacados del IGN
Codigos y nombres oficiales de los municipios (SPAIN). Sacados del INE.
8.124 municipios + 81 condominio + 6 islitas alrededor de La Gomera = 8.211 rows
}
\examples{
\dontrun{
 df <- IGN_mun_17s
}

}
\keyword{datasets}
