% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGN_data_lindes.R
\docType{data}
\name{IGN_mun_17s}
\alias{IGN_mun_17s}
\title{LINDES del IGN para Municipios (SPAIN) (short version)}
\format{A data frame with 8211 rows and 7 variables:
\itemize{
\item INECodMuni: Codigo INE municipio (5 digitos (Pancrudo = 44177))
\item NATCODE: Codigo IGN (Pancrudo = 34024444177)
\item NombreMuni: Nombre municipio (INE)
\item INECodProv: Codigo provincial(INE)(Pancrudo = 44)
\item INECodCCAA: Codigo CCAA(INE) (Pancrudo = 02)
\item NombreProv: Nombre provincia (INE)
\item NombreCCAA: Nombre de la C.A (INE)
\item geometry: la geometria
}}
\source{
\url{http://centrodedescargas.cnig.es/CentroDescargas/index.jsp}
}
\usage{
IGN_mun_17s
}
\description{
Poligonos/geometrias/shapes de los municipios  CC.AA. (SPAIN). Sacados del IGN
Codigos y nombres oficiales de los municipios (SPAIN). Sacados del INE.
8.124 municipios + 81 condominio + 6 islitas alrededor de La Gomera = 8.211 rows
}
\examples{
\dontrun{
 df <- IGN_mun_17s
}

}
\keyword{datasets}
