% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGN_data_lindes.R
\docType{data}
\name{IGN_prov_17}
\alias{IGN_prov_17}
\title{LINDES del IGN para PROVINCIAS (SPAIN)}
\format{A sf-data frame with 52 rows and 13 variables:
\itemize{
\item INECodProv: Código provincia (INE) (Teruel = 44)
\item NATCODE: Code IGN (Teruel = 34024400000)
\item NAMEUNIT: Nombre de la provincia (IGN)
\item NombreProv: Nombre de la provincia (INE)
\item NombreCCAA: Nombre C.A (INE)
\item INSPIREID: Teruel = ES.IGN.SIGLIM34024400000
\item COUNTRY: ES = Spain
\item NATLEVEL: http://inspire.ec.europa.eu/codelist/AdministrativeHierarchyLevel/3rdOrder
\item CODNUT1: (Teruel = ES2)
\item CODNUT2: (Teruel = ES24)
\item CODNUT3: NA
\item INECodCCAA: Código de la C.A (INE)
\item geometry: la geometria
}}
\source{
\url{http://centrodedescargas.cnig.es/CentroDescargas/index.jsp}
}
\usage{
IGN_prov_17
}
\description{
Poligonos/geometrias/shapes de las provincias  CC.AA. (SPAIN). Sacados del IGN
Codigos y nombres oficiales de las provincias y CCAA (SPAIN). Sacados del INE.
52 provincias (51 = Ceuta y 52 = Melilla).
}
\examples{
\dontrun{
 df <- IGN_prov_17
}

}
\keyword{datasets}
