% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGN_data_lindes.R
\docType{data}
\name{IGN_prov_17s}
\alias{IGN_prov_17s}
\title{LINDES del IGN para PROVINCIAS (SPAIN) (Short version)}
\format{A sf-data frame with 52 rows and 6 variables:
\itemize{
\item INECodProv: Código provincia (INE) (Teruel = 44)
\item NATCODE: Code IGN (Teruel = 34024400000)
\item NombreProv: Nombre de la provincia (INE)
\item NombreCCAA: Nombre C.A (INE)
\item INECodCCAA: Código de la C.A (INE)
\item geometry: la geometria
}}
\source{
\url{http://centrodedescargas.cnig.es/CentroDescargas/index.jsp}
}
\usage{
IGN_prov_17s
}
\description{
Poligonos/geometrias/shapes de las provincias  CC.AA. (SPAIN). Sacados del IGN
Codigos y nombres oficiales de las provincias y CCAA (SPAIN). Sacados del INE.
52 provincias (51 = Ceuta y 52 = Melilla).
}
\examples{
\dontrun{
 df <- IGN_prov_17
}

}
\keyword{datasets}
