% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanishRshapes-package.R
\docType{package}
\name{spanishRshapes-package}
\alias{spanishRshapes}
\alias{spanishRshapes-package}
\title{spanishRshapes: Provides some Spanish Boundary data (shapes) as 'sf'-Ready Data Frames}
\description{
spanishRshapes package provides Spanish Boundary data (shapes) as 'sf'-Ready Data Frames. Boundary Data comes from
Instituto Goegráfico Nacional (IGN) and some codes from INE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/perezp44/spanishRshapes}
  \item Report bugs at \url{https://github.com/perezp44/spanishRshapes/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Perez \email{pedro.j.perez@uv.es}

}
\keyword{internal}
